refactor: extract optional and nullable
