+++ packages/pure-parse/src/parse/index.ts
@@ -6,3 +6,7 @@ export * from './primitives'
 export * from './fallback'
 export * from './json'
 export { literal } from './literal'
+export { optionalNullable } from './optional'
+export { undefineable } from './optional'
+export { nullable } from './optional'
+export { optional } from './optional'
