+++ packages/pure-parse/src/parse/types.test.ts
@@ -1,10 +1,11 @@
 import { describe, expect, it, test } from 'vitest'
 import { array } from './array'
 import { object } from './object'
-import { optional, union } from './union'
+import { union } from './union'
 import { parseNumber, parseString } from './primitives'
 import { fallback } from './fallback'
 import { literal } from './literal'
+import { optional } from './optional'
 
 describe('parsing', () => {
   describe('some use cases', () => {
